fix: constant warnings in cache erased
The scheduler clutters the log with warnings due to
various erronous behaviours in the caching.

When we add a Task to the scheduler node cache we explicitly
omit the tasks in the event_handler that are Terminated, but
we try to delete them afterwards. This results in these log lines:
W0925 07:16:56.564363    1698 node_info.go:466] failed to find task <default/nginx2-67755588bf-pkfhp> on host <k3d-volcano-dev-agent-2>

Adding the logic to not remove the tasks that are terminated already
fixes the issue.

The jobId null message was quite intimidating too, although its
pretty common that there are workloads which are not scheduled by
volcano so it's popping up. Modifing it to be an Info message.

Signed-off-by: Hajnal Máté <hajnalmt@gmail.com>